added new effect

+++ Utils/TransistionLibrary.py
@@ -12,7 +12,7 @@ def LinearTransistion(c1, c2, N, options=None):
     if type(c1) == type([]):
         C_Gen = np.zeros((N, len(c1)))
         for i in range(len(c1)):
-            C_Gen[:, i] = np.linspace(c1[i], c2[i], N).astype(np.uint8)
+            C_Gen[:, i] = np.linspace(c1[i], c2[i], N).astype(int)
     else:
-        C_Gen = np.linspace(c1, c2, N).astype(np.uint8)
+        C_Gen = np.linspace(c1, c2, N).astype(int)
     return list(C_Gen)
